% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC.R
\name{DIC}
\alias{DIC}
\title{DIC}
\usage{
DIC(model)
}
\arguments{
\item{model}{Object from \code{\link{GEVmodel}}}
}
\value{
DIC value
}
\description{
This function calculates the DIC of the models retuned by the 
\code{\link{GEVmodel}} function.
}
\references{
Castillo-Mateo J, Asín J, Cebrián AC, Mateo-Lázaro J, Abaurrea J (2023). 
“Bayesian variable selection in generalized extreme value regression: Modeling annual maximum temperature.” 
\emph{Mathematics}, \strong{11}(3), 759. \doi{10.3390/math11030759}.
}
