% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEV.R
\name{GEVmodel}
\alias{GEVmodel}
\title{GEV model with SSVS}
\usage{
GEVmodel(
  Y,
  X = NULL,
  n.sims = 1e+05,
  inits = NULL,
  prior = c(100^2, 100^2, 3, 1/100^2, 100^2, 0.5)
)
}
\arguments{
\item{Y}{VECTOR of data}

\item{X}{MATRIX of covariates (design matrix without intercept)}

\item{n.sims}{Number of simulations}

\item{inits}{Initial values (\eqn{\sigma} transformed scale \eqn{(-\infty,\infty)})}

\item{prior}{Vector. Variance of the zero-mean normal prior for 
(1) \eqn{\beta_0}, (2) \eqn{\log \sigma}, (3) \eqn{\xi}, and value of
(4) \eqn{\tau_i^2}, (5) \eqn{c_i^2 \tau_i^2}, and (6) \eqn{p_i}}
}
\value{
A \code{"GEVmodel"} list with elements:
  \item{params}{Matrix where rows are simulations and cols are parameters
  \deqn{\beta_0,\sigma,\xi,\beta_1,\ldots,\beta_p,\gamma_1,\ldots,\gamma_p}}
  \item{\code{y}}{Data fitted}
  \item{\code{x}}{Covariates}
}
\description{
This function fits the GEV model
}
