% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEV.R
\name{GEVmodel}
\alias{GEVmodel}
\title{GEV-SSVS model}
\usage{
GEVmodel(
  Y,
  X = NULL,
  inits = NULL,
  const = 1,
  prior = c(10^2, 10^2, 1, 0.05^2, 5^2, 0.5),
  n.sims = 1e+05,
  n.thin = 1,
  n.burnin = 10000,
  n.report = 10000
)
}
\arguments{
\item{Y}{VECTOR of responses}

\item{X}{MATRIX of covariates (design matrix without intercept)}

\item{inits}{Initial values (\eqn{\sigma} in log scale 
\eqn{(-\infty,\infty)})}

\item{const}{Constant to scale the adaptive variance of the proposal 
distribution in the MCMC}

\item{prior}{Vector. Variance of the zero-mean normal prior for 
(1) \eqn{\beta_0}, (2) \eqn{\log \sigma}, (3) \eqn{\xi}, and value of
(4) \eqn{\tau_i^2}, (5) \eqn{c_i^2 \tau_i^2}, and (6) \eqn{p_i}}

\item{n.sims, n.thin, n.burnin, n.report}{(i) Number of iterations not 
discarded. (ii) Thinning rate. (iii) Number of iterations discarded at the 
beginning. (iv) Report the number of iterations rate.}
}
\value{
A \code{"GEVmodel"} list with elements:
  \item{params}{Matrix where rows are MCMC simulations and cols are 
    parameters:
  \deqn{\beta_0,\sigma,\xi,\beta_1,\ldots,\beta_p,\gamma_1,\ldots,\gamma_p}}
  \item{\code{y}}{Data fitted}
  \item{\code{x}}{Covariates}
}
\description{
This function fits the GEV-SSVS model in a Bayesian framework using MCMC.
}
\details{
The fitting algorithm is an adaptive Metropolis-within-Gibbs MCMC algorithm.
It has a preset adaptation period. In addition, a burn-in period can be 
established in the arguments.
}
\note{
Set the prior for \eqn{p_i} to \eqn{1} for a usual GEV model, then the 5th 
element of \code{prior} is the prior variance of the parameters associated 
with the covariates.
}
\references{
Castillo-Mateo J, Asín J, Cebrián AC, Mateo-Lázaro J, Abaurrea J (2023). 
“Bayesian variable selection in generalized extreme value regression: Modeling annual maximum temperature.” 
\emph{Mathematics}, \strong{11}(3), 759. \doi{10.3390/math11030759}.
}
