% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEV.R
\name{GEVmodel}
\alias{GEVmodel}
\title{GEV model with SSVS}
\usage{
GEVmodel(
  Y,
  X = NULL,
  inits = NULL,
  prior = c(10^2, 10^2, 1, 0.023^2, 0.23^2, 0.5),
  n.sims = 1e+05,
  n.thin = 1,
  n.burnin = 10000,
  n.report = 10000
)
}
\arguments{
\item{Y}{VECTOR of data}

\item{X}{MATRIX of covariates (design matrix without intercept)}

\item{inits}{Initial values (\eqn{\sigma} transformed scale \eqn{(-\infty,\infty)})}

\item{prior}{Vector. Variance of the zero-mean normal prior for 
(1) \eqn{\beta_0}, (2) \eqn{\log \sigma}, (3) \eqn{\xi}, and value of
(4) \eqn{\tau_i^2}, (5) \eqn{c_i^2 \tau_i^2}, and (6) \eqn{p_i}}

\item{n.sims, n.thin, n.burnin, n.report}{(i) Number of iterations. (ii) 
Thinning rate. (iii) Number of iterations discarded at the beginning. (iv)
Report the number of iterations rate.}
}
\value{
A \code{"GEVmodel"} list with elements:
  \item{params}{Matrix where rows are simulations and cols are parameters
  \deqn{\beta_0,\sigma,\xi,\beta_1,\ldots,\beta_p,\gamma_1,\ldots,\gamma_p}}
  \item{\code{y}}{Data fitted}
  \item{\code{x}}{Covariates}
}
\description{
This function fits the GEV model with SSVS.
Set the prior for \eqn{p_i} to 1 for a usual GEV model.
}
