% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEVSSVS.R
\docType{package}
\name{GEVSSVS}
\alias{GEVSSVS}
\title{GEV-SSVS regression}
\description{
Generalized Extreme Value (GEV) regression with Stochastic Search Variable 
Selection (SSVS).
}
\author{
Jorge Castillo-mateo <jorgecastillomateo@gmail.com>
}
